requirements.txt
# Core
numpy>=1.25.0
tqdm>=4.65.0

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.9.0

# Sentence Transformers for embeddings
sentence-transformers>=2.2.2
transformers>=4.40.0
torch>=2.1.0

# Pinecone with gRPC & asyncio support
pinecone-client[grpc,asyncio]>=2.2.0

# OpenAI API (OpenRouter)
openai>=1.33.0

# Tavily (web search)
tavily>=0.1.3

# dotenv for environment variables
python-dotenv>=1.0.0

# CrossEncoder for reranker
scikit-learn>=1.3.0
